requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
ciso8601==2.3.1
colorama==0.4.6
colorlog==6.9.0
frozenlist==1.5.0
hikari==2.1.0
hikari-crescent==1.1.0
idna==3.10
multidict==6.1.0
orjson==3.10.11
propcache==0.2.0
python-dotenv==1.0.1
sigparse==3.0.0
typing_extensions==4.12.2
yarl==1.17.1
